Moved token icons
