+++ src/components/icons/assets/tokens/BCH.tsx
