+++ src/components/icons/assets/tokens/BTC.tsx
