+++ src/components/icons/assets/tokens/DOGE.tsx
