+++ src/components/icons/assets/tokens/LTC.tsx
