+++ src/components/icons/assets/tokens/USDC.tsx
