+++ src/components/icons/assets/tokens/USDT.tsx
