+++ src/components/icons/assets/tokens/_Default.tsx
