+++ src/components/icons/assets/tokens/dAAPL.tsx
