+++ src/components/icons/assets/tokens/dAMD.tsx
