+++ src/components/icons/assets/tokens/dBABA.tsx
