+++ src/components/icons/assets/tokens/dBCH.tsx
