+++ src/components/icons/assets/tokens/dGLD.tsx
