+++ src/components/icons/assets/tokens/dGME.tsx
