+++ src/components/icons/assets/tokens/dGOOG.tsx
