+++ src/components/icons/assets/tokens/dLTC.tsx
