+++ src/components/icons/assets/tokens/dPLTR.tsx
