+++ src/components/icons/assets/tokens/dQQQ.tsx
