+++ src/components/icons/assets/tokens/dSLV.tsx
