+++ src/components/icons/assets/tokens/dSPY.tsx
