+++ src/components/icons/assets/tokens/dTSLA.tsx
