+++ src/components/icons/assets/tokens/dVNQ.tsx
