+++ src/components/vaults/common/VaultAuctionDetails.tsx
@@ -5,7 +5,7 @@ import BigNumber from 'bignumber.js'
 import { VaultTokenSymbols } from '@components/vaults/common/VaultTokenSymbols'
 import { CalculateCollateralsValue } from '../../../utils/vaults/CalculateCollateralsValue'
 import { LoanVaultLiquidated, LoanVaultLiquidationBatch } from '@defichain/whale-api-client/dist/api/loan'
-import { getAssetIcon } from '@components/icons/assets'
+import { getAssetIcon } from '@components/icons/assets/tokens'
 import { VaultDetailsListItem } from '@components/vaults/common/VaultDetailsListItem'
 import { BidAmountValue } from '@components/auctions/commons/BidAmountValue'
 import React from 'react'
